Removing ffmpeg packages
