--- app/static/main.css
@@ -1,227 +1,0 @@
-﻿.container {
-  position: relative;
-  width: 100%;
-  max-width: 1440px;
-  margin: 0 auto;
-  padding: 0 20px;
-  box-sizing: border-box;
-}
-
-.column,
-.columns {
-  width: 100%;
-  float: left;
-  box-sizing: border-box;
-}
-
-.row {
-  padding: 10px;
-
-    padding-top: 10vh;
-}
-
-@media (min-width: 400px) {
-  .container {
-    width: 85%;
-    padding: 0;
-  }
-}
-@media (min-width: 550px) {
-  .container {
-    width: 80%;
-  }
-
-  .column:first-child,
-.columns:first-child {
-    margin-left: 0;
-  }
-
-  .column,
-.columns {
-    margin-left: 4%;
-  }
-
-  .one.column,
-.one.columns {
-    width: 4.6666666667%;
-  }
-
-  .two.columns {
-    width: 13.3333333333%;
-  }
-
-  .three.columns {
-    width: 22%;
-  }
-
-  .four.columns {
-    width: 30.6666666667%;
-  }
-
-  .five.columns {
-    width: 39.3333333333%;
-  }
-
-  .six.columns {
-    width: 48%;
-  }
-
-  .seven.columns {
-    width: 56.6666666667%;
-  }
-
-  .eight.columns {
-    width: 65.3333333333%;
-  }
-
-  .nine.columns {
-    width: 74%;
-  }
-
-  .ten.columns {
-    width: 82.6666666667%;
-  }
-
-  .eleven.columns {
-    width: 91.3333333333%;
-  }
-
-  .twelve.columns {
-    width: 100%;
-    margin-left: 0;
-  }
-
-  .one-third.column {
-    width: 30.6666666667%;
-  }
-
-  .two-thirds.column {
-    width: 65.3333333333%;
-  }
-
-  .one-half.column {
-    width: 48%;
-  }
-
-.container:after,
-.row:after,
-.u-cf {
-  content: "";
-  display: table;
-  clear: both;
-}
-
-
-
-h1 { font-family: "Segoe UI", Frutiger, "Frutiger Linotype", "Dejavu Sans", "Helvetica Neue", Arial, sans-serif; font-size: 24px; font-style: normal; font-variant: normal; font-weight: 700; line-height: 26.4px; } h3 { font-family: "Segoe UI", Frutiger, "Frutiger Linotype", "Dejavu Sans", "Helvetica Neue", Arial, sans-serif; font-size: 14px; font-style: normal; font-variant: normal; font-weight: 700; line-height: 15.4px; } a, p, input { font-family: "Segoe UI", Frutiger, "Frutiger Linotype", "Dejavu Sans", "Helvetica Neue", Arial, sans-serif; font-size: 14px; font-style: normal; font-variant: normal; font-weight: 400; line-height: 20px; } blockquote { font-family: "Segoe UI", Frutiger, "Frutiger Linotype", "Dejavu Sans", "Helvetica Neue", Arial, sans-serif; font-size: 21px; font-style: normal; font-variant: normal; font-weight: 400; line-height: 30px; } pre { font-family: "Segoe UI", Frutiger, "Frutiger Linotype", "Dejavu Sans", "Helvetica Neue", Arial, sans-serif; font-size: 13px; font-style: normal; font-variant: normal; font-weight: 400; line-height: 18.5714px; }
-
-div.header {
-  background-color: #191919;
-  height: 98vh;
-  width: 100%;
-}
-
-body {
-  margin: 0px;
-  color: white;
-  text-align: center;
-  overflow-x: hidden;
-}
-
-img {
-  align-self: center;
-}
-
-h1 {
-  font-size: 5rem;
-}
-
-h3 {
-  font-size: 1.5rem;
-}
-
-img#hum {
-  height: 9vh;
-}
-
-.jumbotron {
-  position: relative;
-  display: block;
-  top: 20%;
-}
-
-form {
-  width: 60vh;
-  margin: auto;
-  margin-top: 40px;
-}
-
-input {
-  display: inline;
-  position: relative;
-}
-
-span {
-  text-align: center;
-}
-
-#url {
-  width: 50vh;
-  height: 6vh;
-  background-color: black;
-  opacity: .5;
-  border: none;
-  float: left;
-  color: white;
-  padding-left: 20px;
-}
-
-#submit {
-  width: 10vh;
-  height: 6vh;
-  background-image: url('Symbol 3 – 1.png');
-  border: none;
-  color: white;
-  font-weight: 600;
-  font-size: 15px;
-}
-
-#git {
-  text-decoration: none;
-  color: white;
-  font-size: 1.2rem;
-  margin-top: 3rem;
-}
-
-.content {
-  background-image: url('Symbol 3 – 1@2x.png');
-  background-repeat: space;
-  height: 100%;
-  width: 100%;
-  padding-bottom: 270vh;
-}
-
-.videos {
-  padding-top: 10vh;
-}
-
-.video {
-  display: block;
-  box-shadow: 0px 0px 50px black;
-}
-
-.title img {
-  margin-top: 5vh;
-}
-
-video {
-  display: block;
-  width: 100%;
-  background-color: #191919;
-}
-
-.title {
-	display:block;
-	background-color: #191919;
-  box-shadow: 0px 0px 50px black;
-}
