--- app/templates/theonlyhtmlfileweneed.html
@@ -1,119 +1,0 @@
-  <html>
-<head>
-  {% if title %}
-  <title>{{ title }}</title>
-  {% else %}
-  <title>Song Flong</title>
-  {% endif %}
-  <link rel= "stylesheet" type= "text/css" href= "{{ url_for('static',filename='main.css') }}">
-</head>
-<body>
-  <p id="log"></p>
-<div class="wrapper">
-  <div class="header">
-    <div class="container jumbotron">
-      <img src="https://hackumass.com/assets/img/logo.png" alt="hackumass" id="hum">
-      <h1>Song Flong</h1>
-      <h3>Can a music video still work with a different song?</h3>
-
-      <form action="" method="post">
-
-        {{ form.csrf_token }}
-          <div>
-                      {% with messages = get_flashed_messages() %}
-                      {% if messages %}
-                      <ul>
-                          {% for message in messages %}
-                          <li>{{ message }}</li>
-                          {% endfor %}
-                      </ul>
-                      {% endif %}
-                      {% endwith %}
-              {{ form.url(size=32, placeholder="YouTube URL...") }}
-
-          <p>{{ form.submit() }}</p>
-        </div>
-        {% for error in form.url.errors %}
-              <p style="color: #C20669;">{{ error }}</p>
-{% endfor %}
-      </form>
-
-      <a id="git" href="https://github.com/skylers27/Sound-Repo-Thing">Source Code</a>
-
-    </div>
-  </div>
-
-  <div class="content" style="clear:both">
-    <div class="videos container">
-
-      <div class="row">
-
-        <div class="video two-thirds column">
-          <video id="video0" controls>
-          <source src="{{ videoData[0].final }}" type="video/mp4"></video>
-        </div>
-        <div class="title one-third column" height="document.getElementsByClassName('row')[0].offsetWidth;">
-                    <img id="art0" src=" {{ videoData[0].art | default('https://s3.amazonaws.com/assets.jog.fm/images/missing-600x600.png') }}" alt="Album art" height="240px" width="240px">
-            <h3>{{ videoData[0].title }}</h3>
-            <p>{{ videoData[0].artist }}</p>
-        </div>
-      </div>
-      <div class="row">
-        <div class="title one-third column">
-          <img id="art1"  src=" {{ videoData[1].art | default('https://s3.amazonaws.com/assets.jog.fm/images/missing-600x600.png') }}" alt="Album art" height="240px" width="240px">
-<h3>{{ videoData[1].title }}</h3>
-<p>{{ videoData[1].artist }}</p>
-        </div>
-
-          <div class="video two-thirds column">
-            <video id="video1" controls>
-            <source src="{{ videoData[1].final }}"  type="video/mp4"></video>
-          </div>
-      </div>
-      <div class="row">
-
-        <div class="video two-thirds column">
-          <video id="video2" controls>
-          <source src="{{ videoData[2].final }}"  type="video/mp4"></video>
-        </div>
-        <div class="title one-third column">
-                    <img id="art2"  src=" {{ videoData[2].art | default('https://s3.amazonaws.com/assets.jog.fm/images/missing-600x600.png') }}" alt="Album art" height="240px" width="240px">
-<h3>{{ videoData[2].title }}</h3>
-<p>{{ videoData[2].artist }}</p>
-        </div>
-      </div>
-      <div class="row">
-        <div class="title one-third column">
-                    <img id="art3"  src=" {{ videoData[3].art | default('https://s3.amazonaws.com/assets.jog.fm/images/missing-600x600.png') }}" alt="Album art" height="240px" width="240px">
-<h3>{{ videoData[3].title }}</h3>
-<p>{{ videoData[3].artist }}</p>
-        </div>
-
-          <div class="video two-thirds column">
-            <video  id="video3" controls>
-            <source src="{{ videoData[3].final }}"  type="video/mp4"></video>
-          </div>
-      </div>
-      <div class="row">
-
-        <div class="video two-thirds column">
-          <video id="video4" controls>
-          <source src="{{ videoData[4].final }}"  type="video/mp4"></video>
-        </div>
-        <div class="title one-third column">
-                    <img id="art4"  src=" {{ videoData[4].art | default('https://s3.amazonaws.com/assets.jog.fm/images/missing-600x600.png') }}" alt="Album art" height="240px" width="240px">
-<h3>{{ videoData[4].title }}</h3>
-<p>{{ videoData[4].artist }}</p>
-        </div>
-      </div>
-    </div>
-  </div>
-</div>
-<script src="{{ url_for('static',filename='ResizeSensor.js') }}"></script>
-<script src="{{ url_for('static',filename='ElementQueries.js') }}"></script>
-<script src="{{ url_for('static',filename='main.js') }}"></script>
-<script type="text/javascript" src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.3.1/jquery.min.js"></script>
-<script type="text/javascript" src="//cdnjs.cloudflare.com/ajax/libs/socket.io/1.3.6/socket.io.min.js"></script>
-<script src="{{ url_for('static',filename='videoUpdate.js') }}"></script>
-</body>
-</html>
