+++ requirements.txt
@@ -1,3 +1,4 @@
+arrow==0.15.2
 beautifulsoup4==4.8.0
 bs4==0.0.1
 cachetools==3.1.1
@@ -5,29 +6,29 @@ certifi==2019.9.11
 chardet==3.0.4
 Click==7.0
 decorator==4.4.0
+ffmpeg-python==0.2.0
 Flask==1.1.1
 Flask-SocketIO==4.2.1
 Flask-WTF==0.14.2
+future==0.17.1
 idna==2.8
-imageio==2.5.0
-imageio-ffmpeg==0.3.0
 itsdangerous==1.1.0
 Jinja2==2.10.1
 lxml==4.4.1
 MarkupSafe==1.1.1
-numpy==1.17.2
 oauth2client==4.1.3
 Pillow==6.1.0
-pkg-resources==0.0.0
 proglog==0.1.9
 pyasn1==0.4.7
 pyasn1-modules==0.2.6
+python-dateutil==2.8.0
 python-engineio==3.9.3
 python-socketio==4.3.1
 pytube==9.5.2
 redis==3.3.8
 requests==2.22.0
 rq==1.1.0
+rq-dashboard==0.5.3
 rsa==4.0
 six==1.12.0
 soupsieve==1.9.3
