+++ run.py
@@ -1,4 +1,4 @@
-from app2.app import app
+from app.app import app
 
 
 app.run(host='0.0.0.0', port=8080, debug=True)
